@@ -1,115 +1,46 @@
 {
-  
-  "name": "cabinbooking",
-  
+  "name": "react-firebase-auth",
   "version": "0.1.0",
-  
   "private": true,
-  
   "dependencies": {
-  
-    
-  "@testing-library/jest-dom": "^4.2.4",
-  
-    
-  "@testing-library/react": "^9.5.0",
-  
-    
-  "@testing-library/user-event": "^7.2.1",
-  
-    
-  "firebase": "^9.9.4",
-  
-    
-  "react": "^16.13.1",
-  
-    
-  "react-datetime-picker": "^3.5.0",
-  
-    
-  "react-dom": "^16.13.1",
-  
-    
-  "react-router-dom": "^5.2.0",
-  
-    
-  "react-scripts": "3.4.1"
-  
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^13.5.0",
+    "firebase": "^9.9.4",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.5.0",
+    "react-datetime-picker": "^3.5.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
+    "redux": "^4.2.0",
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.4.1",
+    "web-vitals": "^2.1.4"
   },
-  
   "scripts": {
-  
-    
-  "predeploy": "npm run build",
-  
-    
-  "deploy": "gh-pages -d build",
-  
-    
-  "start": "react-scripts start",
-  
-    
-  "build": "react-scripts build",
-  
-    
-  "test": "react-scripts test",
-  
-    
-  "eject": "react-scripts eject"
-  
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
   },
-  
   "eslintConfig": {
-  
-    
-  "extends": "react-app"
-  
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
-  
   "browserslist": {
-  
-    
-  "production": [
-  
-    
-    
-  ">0.2%",
-  
-    
-    
-  "not dead",
-  
-    
-    
-  "not op_mini all"
-  
-    
-  ],
-  
-    
-  "development": [
-  
-    
-    
-  "last 1 chrome version",
-  
-    
-    
-  "last 1 firefox version",
-  
-    
-    
-  "last 1 safari version"
-  
-    
-  ]
-  
-  },
-  
-  "devDependencies": {
-  
-    
-  "gh-pages": "^3.1.0"
-  
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
